parser/htmlparser/tests/mochitest/html5lib_tree_construction/adoption02.dat

#data
<b>1<i>2<p>3</b>4
#errors
(1,3): expected-doctype-but-got-start-tag
(1,16): adoption-agency-1.3
(1,17): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "1"
|       <i>
|         "2"
|     <i>
|       <p>
|         <b>
|           "3"
|         "4"

#data
<a><div><style></style><address><a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,35): unexpected-start-tag-implies-end-tag
(1,35): adoption-agency-1.3
(1,35): adoption-agency-1.3
(1,35): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <div>
|       <a>
|         <style>
|       <address>
|         <a>
|         <a>

#data
<b><i><a><s><tt><div></b>first</b></div></tt></s></a>second</i>
#errors
3: Start tag seen without seeing a doctype first. Expected "<!DOCTYPE html>".
25: End tag "b" violates nesting rules.
34: Stray end tag "b".
63: Stray end tag "i".
#document
| <html>
|   <head>
|   <body>
|     <b>
|       <i>
|         <a>
|           <s>
|             <tt>
|     <a>
|       <s>
|         <tt>
|           <div>
|             <b>
|             "first"
|     "second"
